pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.3",
    "google-genai>=1.32.0",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
]
